refactor(auth): move Register into auth feature
