+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css"
 import { Routes, Route, Navigate } from "react-router-dom"
-import { Register } from "./pages/Register"
+import { Register } from "./features/auth/components/Register"
 import { Login } from "./features/auth/components/Login"
 import { TaskList } from "./features/tasks/components/TaskList"
